requirements.txt
Flask
flask-cors
python-dotenv
requests
pdfplumber
gunicorn